# api/openapi.yaml
openapi: 3.1.0
info:
  title: LyraNest Community API
  version: 0.3.0
  description: Authenticated API for library discovery, playback, management, normalized sidecar lyrics, and persistent per-user listening data.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      security: []
      operationId: registerFirstAdministrator
      summary: Register the first and only bootstrap administrator
      description: Available only while the user database is empty. Registration closes atomically after the first successful request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "201":
          description: Administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Registration is closed because a user already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      security: []
      operationId: login
      summary: Create a persistent server-side session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "200":
          description: Login succeeded
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/tracks:
    get:
      operationId: listTracks
      summary: List scanned tracks
      responses:
        "200":
          description: Scanned tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/tracks/{trackId}/artwork:
    parameters:
      - name: trackId
        in: path
        required: true
        schema:
          type: string
    get:
      security: []
      operationId: getTrackArtwork
      summary: Get track artwork
      description: Returns the preferred sidecar artwork or falls back to embedded artwork from the audio file.
      responses:
        "200":
          $ref: "#/components/responses/Artwork"
        "304":
          description: Artwork has not changed for the supplied conditional request headers.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      security: []
      operationId: headTrackArtwork
      summary: Inspect track artwork
      responses:
        "200":
          $ref: "#/components/responses/ArtworkHeaders"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired, or revoked bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request body, candidate signature, provider, or field selection
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Track, audio file, artwork, or lyrics not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Artwork:
      description: Track artwork image
      headers:
        Cache-Control:
          schema:
            type: string
          description: Public artwork cache policy.
        ETag:
          schema:
            type: string
          description: Content hash used for conditional requests.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: Artwork or audio file modification time in HTTP-date format.
        X-Content-Type-Options:
          schema:
            type: string
            const: nosniff
      content:
        image/jpeg:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        image/png:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        image/webp:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
    ArtworkHeaders:
      description: Track artwork headers with no response body
      headers:
        Cache-Control:
          schema:
            type: string
        Content-Type:
          schema:
            type: string
        ETag:
          schema:
            type: string
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: Artwork or audio file modification time in HTTP-date format.
        X-Content-Type-Options:
          schema:
            type: string
            const: nosniff
...
  schemas:
...
    CredentialsRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 1024
          writeOnly: true
    User:
      type: object
      required: [id, username, role, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          const: admin
        created_at:
          type: string
          format: date-time
    SessionResponse:
      type: object
      required: [token, token_type, expires_at, user]
      properties:
        token:
          type: string
          writeOnly: true
        token_type:
          type: string
          const: Bearer
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ArtworkBinary:
      type: string
      format: binary
...
    TrackListResponse:
      type: object
      required: [tracks, total]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
          minimum: 0
...
    Track:
      type: object
      required: [id, title, file_name, extension, size_bytes, modified]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        limit:
          type: integer
          minimum: 1
          maximum: 25
          default: 15
          description: Maximum number of provider candidates to search for.
        file_name:
          type: string
        extension:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        modified:
          type: string
          format: date-time
        artwork_url:
          type: string
          format: uri
          description: Optional explicit artwork URL. Clients may use the track artwork endpoint when omitted.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string